consolidate serialization
