+++ packages/account/src/providers/provider.ts
@@ -66,9 +66,8 @@ import {
 } from './utils';
 import type { RetryOptions } from './utils/auto-retry-fetch';
 import { autoRetryFetch } from './utils/auto-retry-fetch';
-import { deserializeChain } from './utils/chain-info';
 import { assertGqlResponseHasNoErrors } from './utils/handle-gql-error-message';
-import { deserializeNodeInfo } from './utils/node-info';
+import { deserializeChain, deserializeNodeInfo } from './utils/serialization';
 import { validatePaginationArgs } from './utils/validate-pagination-args';
 
 const MAX_RETRIES = 10;
